+++ .check.exs
@@ -1,0 +1,20 @@
+[
+  parallel: false,
+  skipped: false,
+  tools: [
+    {:deps_get, command: "mix deps.get --check-locked"},
+    {:compiler, command: "mix compile --warnings-as-errors"},
+    {:formatter, command: "mix format --check-formatted"},
+    {:unused_deps, command: "mix deps.unlock --check-unused"},
+    {:credo, command: "mix credo --strict"},
+    {:ex_unit, false},
+    {:coverage, command: "mix coveralls", env: %{"MIX_ENV" => "test"}},
+    {:hex_audit, command: "mix hex.audit"},
+    {:mix_audit, command: "mix deps.audit"},
+    {:doctor, command: "mix doctor"},
+    {:dialyzer, command: "mix dialyzer"},
+    {:ex_doc, command: "mix docs --warnings-as-errors"},
+    {:boundary, command: "bin/check-boundary"},
+    {:package, command: "bin/check-package"}
+  ]
+]